Improve exposition library coherency.
The Go library assumes 64-bit precision with its telemetric
primitives, and the Java library was built with the same intention
but the rapidity of its being built meant that it used 32-bit floats
on accident.  This has been updated in this pull request.

Secondarily, for additional consistency with the Go library, the
reported metric percentile values are stringified with the same
precision formatting rules as the Go ones.  It was possible for
both system locale to introduce incoherency with the formatting.

+++ client/src/main/java/io/prometheus/client/metrics/histogram/Bucket.java
@@ -18,11 +18,11 @@ package io.prometheus.client.metrics.histogram;
  * @author matt.proud@gmail.com (Matt T. Proud)
  */
 public interface Bucket {
-  public void add(final float value);
+  public void add(final double value);
 
   public int observations();
 
   public void reset();
 
-  public float valueForIndex(final int index);
+  public double valueForIndex(final int index);
 }
